problem-supplier-verification-ai-engine/pyproject.toml
[project]
name = "problem-supplier-verification-ai-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.28.0",
    "celery>=5.3.6",
    "redis>=5.0.3",
    "google-generativeai>=0.4.0",
    "langchain-google-genai>=1.0.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.28",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
    "duckduckgo-search>=5.0.0",
    "google-genai>=1.70.0",
]
